+++ Editor/Utilities/Shortcuts.cs
@@ -1,0 +1,36 @@
+﻿namespace Ikonoclast.Common.Editor
+{
+    public static class Shortcuts
+    {
+        public const string alt = "&";
+        public const string ctrl = "%";
+        public const string shift = "#";
+        public const string no_mod = "_";
+
+        public const string a = "a";
+        public const string b = "b";
+        public const string c = "c";
+        public const string d = "d";
+        public const string e = "e";
+        public const string f = "f";
+        public const string g = "g";
+        public const string h = "h";
+        public const string i = "i";
+        public const string j = "j";
+        public const string k = "k";
+        public const string m = "m";
+        public const string n = "n";
+        public const string o = "o";
+        public const string p = "p";
+        public const string q = "q";
+        public const string r = "r";
+        public const string s = "s";
+        public const string t = "t";
+        public const string u = "u";
+        public const string v = "v";
+        public const string w = "w";
+        public const string x = "x";
+        public const string y = "y";
+        public const string z = "z";
+    }
+}