+++ Runtime/Constants/Values.cs
@@ -1,0 +1,19 @@
+﻿namespace Ikonoclast.Common
+{
+    public static class Values
+    {
+        public const char
+            FloatPrefix = 'f',
+            StringPrefix = 's',
+            IntegerPrefix = 'i',
+            BooleanPrefix = 'b',
+            Vector2Prefix = 'v';
+
+        public const string
+            FloatDefault = "0",
+            StringDefault = "",
+            IntegerDefault = "0",
+            BooleanDefault = "False",
+            Vector2Default = "(0, 0)";
+    }
+}